TSQR: Move DistTsqr test code into test directory
