+++ packages/tpetra/tsqr/test/Tsqr_TestDistTsqr.cpp
@@ -37,8 +37,14 @@
 // ************************************************************************
 //@HEADER
 
-#include "Tsqr_ConfigDefs.hpp"
+#include "Tsqr_Random_NormalGenerator.hpp"
+#include "Tsqr_generateStack.hpp"
+#include "Tsqr_DistTsqr.hpp"
+#include "Tsqr_GlobalTimeStats.hpp"
+#include "Tsqr_GlobalVerify.hpp"
+#include "Tsqr_printGlobalMatrix.hpp"
 
+#include "Tsqr_TeuchosMessenger.hpp"
 #ifdef HAVE_MPI
 #  include "Teuchos_GlobalMPISession.hpp"
 #  include "Teuchos_oblackholestream.hpp"
@@ -50,142 +56,841 @@
 #include "Teuchos_Time.hpp"
 #include "Teuchos_StandardCatchMacros.hpp"
 
-#include "Tsqr_ParTest.hpp"
-#include "Tsqr_TeuchosMessenger.hpp"
-
+#include <algorithm>
 #ifdef HAVE_TPETRATSQR_COMPLEX
 #  include <complex>
 #endif // HAVE_TPETRATSQR_COMPLEX
-
+#include <iomanip>
+#include <iostream>
 #include <sstream>
 #include <stdexcept>
 #include <vector>
 
-using TSQR::MessengerBase;
-using TSQR::TeuchosMessenger;
-using TSQR::Test::DistTsqrVerifier;
-using TSQR::Test::DistTsqrBenchmarker;
+namespace TSQR {
+  namespace Test {
+    /// \class DistTsqrVerifier
+    /// \brief Generic version of \c DistTsqr accuracy test.
+    template<class Ordinal, class Scalar>
+    class DistTsqrVerifier {
+      TSQR::Random::NormalGenerator<Ordinal, Scalar> gen_;
+      Teuchos::RCP<MessengerBase<Ordinal> > const ordinalComm_;
+      Teuchos::RCP<MessengerBase<Scalar> > const scalarComm_;
+      std::string scalarTypeName_;
+      std::ostream& out_;
+      std::ostream& err_;
+      const bool testFactorExplicit_, testFactorImplicit_;
+      const bool humanReadable_, printMatrices_, debug_;
 
-using Teuchos::RCP;
-using Teuchos::rcp;
-using Teuchos::rcp_implicit_cast;
-using Teuchos::Tuple;
+    public:
+      typedef Ordinal ordinal_type;
+      typedef Scalar scalar_type;
+      typedef typename Teuchos::ScalarTraits<scalar_type>::magnitudeType magnitude_type;
+      typedef typename std::vector<magnitude_type> result_type;
+      typedef Matrix<ordinal_type, scalar_type> matrix_type;
 
+      /// \brief Constructor, with custom seed value
+      ///
+      /// \param scalarComm [in/out] Communicator object over which to
+      ///   test.
+      /// \param seed [in] 4-element vector; the random seed input of
+      ///   TSQR::Random::NormalGenerator (which see, since there are
+      ///   restrictions on the set of valid seeds)
+      /// \param scalarTypeName [in] Human-readable name of the Scalar
+      ///   template type parameter
+      /// \param out [out] Output stream to which to write results
+      /// \param err [out] Output stream to which to write any
+      ///   debugging outputs (if applicable) or errors
+      /// \param testFactorExplicit [in] Whether to test
+      ///   DistTsqr::factorExplicit()
+      /// \param testFactorImplicit [in] Whether to test
+      ///   DistTsqr::factor() and DistTsqr::explicit_Q()
+      /// \param humanReadable [in] Whether printed results should be
+      ///   easy for humans to read (vs. easy for parsers to parse)
+      /// \param debug [in] Whether to write verbose debug output to
+      ///   err
+      DistTsqrVerifier (const Teuchos::RCP<MessengerBase<Ordinal> >& ordinalComm,
+                        const Teuchos::RCP<MessengerBase<Scalar> >& scalarComm,
+                        const std::vector<int>& seed,
+                        const std::string& scalarTypeName,
+                        std::ostream& out,
+                        std::ostream& err,
+                        const bool testFactorExplicit,
+                        const bool testFactorImplicit,
+                        const bool humanReadable,
+                        const bool printMatrices,
+                        const bool debug) :
+        gen_ (seed),
+        ordinalComm_ (ordinalComm),
+        scalarComm_ (scalarComm),
+        scalarTypeName_ (scalarTypeName),
+        out_ (out),
+        err_ (err),
+        testFactorExplicit_ (testFactorExplicit),
+        testFactorImplicit_ (testFactorImplicit),
+        humanReadable_ (humanReadable),
+        printMatrices_ (printMatrices),
+        debug_ (debug)
+      {}
 
-template< class Ordinal, class Scalar >
-class MessengerPairMaker {
-  public:
-    typedef int ordinal_type;
-    typedef Scalar scalar_type;
+      /// \brief Constructor, with default seed value
+      ///
+      /// This constructor sets a default seed (for the pseudorandom
+      /// number generator), which is the same seed (0,0,0,1) each
+      /// time.
+      ///
+      /// \param scalarComm [in/out] Communicator object over which to
+      ///   test.
+      /// \param scalarTypeName [in] Human-readable name of the Scalar
+      ///   template type parameter
+      /// \param out [out] Output stream to which to write results
+      /// \param err [out] Output stream to which to write any
+      ///   debugging outputs (if applicable) or errors
+      /// \param testFactorExplicit [in] Whether to test
+      ///   DistTsqr::factorExplicit()
+      /// \param testFactorImplicit [in] Whether to test
+      ///   DistTsqr::factor() and DistTsqr::explicit_Q()
+      /// \param humanReadable [in] Whether printed results should be
+      ///   easy for humans to read (vs. easy for parsers to parse)
+      /// \param debug [in] Whether to write verbose debug output to
+      ///   err
+      DistTsqrVerifier (const Teuchos::RCP<MessengerBase<Ordinal> >& ordinalComm,
+                        const Teuchos::RCP<MessengerBase<Scalar> >& scalarComm,
+                        const std::string& scalarTypeName,
+                        std::ostream& out,
+                        std::ostream& err,
+                        const bool testFactorExplicit,
+                        const bool testFactorImplicit,
+                        const bool humanReadable,
+                        const bool printMatrices,
+                        const bool debug) :
+        ordinalComm_ (ordinalComm),
+        scalarComm_ (scalarComm),
+        scalarTypeName_ (scalarTypeName),
+        out_ (out),
+        err_ (err),
+        testFactorExplicit_ (testFactorExplicit),
+        testFactorImplicit_ (testFactorImplicit),
+        humanReadable_ (humanReadable),
+        printMatrices_ (printMatrices),
+        debug_ (debug)
+      {}
 
-    typedef std::pair<RCP<MessengerBase<ordinal_type> >, RCP<MessengerBase<scalar_type> > > pair_type;
-
-    static pair_type
-      makePair (const RCP< const Teuchos::Comm<int> >& comm)
+      /// \brief Get seed vector for pseudorandom number generator
+      ///
+      /// Fill seed (changing size of vector as necessary) with the
+      /// seed vector used by the pseudorandom number generator.  You
+      /// can use this to resume the pseudorandom number stream from
+      /// where you last were.
+      void
+      getSeed (std::vector<int>& seed) const
       {
-        RCP<TeuchosMessenger<ordinal_type> > derivedOrdinalComm =
-          rcp (new TeuchosMessenger<ordinal_type> (comm));
-        RCP<MessengerBase<ordinal_type> > ordinalComm =
-          rcp_implicit_cast<MessengerBase<ordinal_type> > (derivedOrdinalComm);
-        RCP<TeuchosMessenger<scalar_type> > derivedScalarComm =
-          rcp (new TeuchosMessenger<scalar_type> (comm));
-        RCP<MessengerBase<scalar_type> > scalarComm =
-          rcp_implicit_cast<MessengerBase<scalar_type> > (derivedScalarComm);
-
-        return std::make_pair (ordinalComm, scalarComm);
+        gen_.getSeed (seed);
       }
-};
 
+      /// \brief Run the DistTsqr accuracy test
+      ///
+      /// \param numCols [in] Number of columns in the matrix to test.
+      ///   Number of rows := (# MPI processors) * ncols.
+      void
+      verify (const Ordinal numCols,
+              const std::string& additionalFieldNames,
+              const std::string& additionalData,
+              const bool printFieldNames)
+      {
+        using std::endl;
+
+        const int myRank = scalarComm_->rank();
+        if (debug_)
+          {
+            scalarComm_->barrier();
+            if (myRank == 0)
+              err_ << "Verifying DistTsqr:" << endl;
+            scalarComm_->barrier();
+          }
+
+        // Generate test problem.
+        Matrix< Ordinal, Scalar > A_local, Q_local, R;
+        testProblem (A_local, Q_local, R, numCols);
+        if (debug_) {
+          scalarComm_->barrier();
+          if (myRank == 0) {
+            err_ << "-- Generated test problem." << endl;
+          }
+          scalarComm_->barrier();
+        }
+
+        // Set up TSQR implementation.
+        DistTsqr<Ordinal, Scalar> par;
+        par.init (scalarComm_);
+        if (debug_) {
+          scalarComm_->barrier();
+          if (myRank == 0) {
+            err_ << "-- DistTsqr object initialized" << endl << endl;
+          }
+        }
+
+        // Whether we've printed field names (i.e., column headers)
+        // yet.  Only matters for non-humanReadable output.
+        bool printedFieldNames = false;
+
+        // Test DistTsqr::factor() and DistTsqr::explicit_Q().
+        if (testFactorImplicit_) {
+          // Factor the matrix A (copied into R, which will be
+          // overwritten on output)
+          typedef typename DistTsqr<Ordinal, Scalar>::FactorOutput
+            factor_output_type;
+          factor_output_type factorOutput = par.factor (R.view());
+          if (debug_) {
+            scalarComm_->barrier();
+            if (myRank == 0) {
+              err_ << "-- Finished DistTsqr::factor" << endl;
+            }
+          }
+          // Compute the explicit Q factor
+          par.explicit_Q (numCols, Q_local.data(), Q_local.stride(1), factorOutput);
+          if (debug_) {
+            scalarComm_->barrier();
+            if (myRank == 0) {
+              err_ << "-- Finished DistTsqr::explicit_Q" << endl;
+            }
+          }
+          // Verify the factorization
+          result_type result =
+            global_verify (numCols, numCols, A_local.data(), A_local.stride(1),
+                           Q_local.data(), Q_local.stride(1), R.data(), R.stride(1),
+                           scalarComm_.get());
+          if (debug_) {
+            scalarComm_->barrier();
+            if (myRank == 0) {
+              err_ << "-- Finished global_verify" << endl;
+            }
+          }
+          reportResults ("DistTsqr", numCols, result,
+                         additionalFieldNames, additionalData,
+                         printFieldNames && (! printedFieldNames));
+          if (printFieldNames && (! printedFieldNames))
+            printedFieldNames = true;
+        }
+
+        // Test DistTsqr::factorExplicit()
+        if (testFactorExplicit_) {
+          // Factor the matrix and compute the explicit Q factor, both
+          // in a single operation.
+          par.factorExplicit (R.view(), Q_local.view());
+          if (debug_) {
+            scalarComm_->barrier();
+            if (myRank == 0) {
+              err_ << "-- Finished DistTsqr::factorExplicit" << endl;
+            }
+          }
+
+          if (printMatrices_) {
+            if (myRank == 0) {
+              err_ << std::endl << "Computed Q factor:" << std::endl;
+            }
+            printGlobalMatrix (err_, Q_local, scalarComm_.get(), ordinalComm_.get());
+            if (myRank == 0) {
+              err_ << std::endl << "Computed R factor:" << std::endl;
+              print_local_matrix (err_, R.extent(0), R.extent(1), R.data(), R.stride(1));
+              err_ << std::endl;
+            }
+          }
+
+          // Verify the factorization
+          result_type result =
+            global_verify (numCols, numCols, A_local.data(), A_local.stride(1),
+                           Q_local.data(), Q_local.stride(1), R.data(), R.stride(1),
+                           scalarComm_.get());
+          if (debug_) {
+            scalarComm_->barrier();
+            if (myRank == 0) {
+              err_ << "-- Finished global_verify" << endl;
+            }
+          }
+          reportResults ("DistTsqrRB", numCols, result,
+                         additionalFieldNames, additionalData,
+                         printFieldNames && (! printedFieldNames));
+          if (printFieldNames && (! printedFieldNames)) {
+            printedFieldNames = true;
+          }
+        }
+      }
+
+    private:
+      /// Report verification results.  Call on ALL MPI processes, not
+      /// just Rank 0.
+      ///
+      /// \param method [in] String to print before reporting results
+      /// \param numCols [in] Number of columns in the matrix tested.
+      /// \param result [in] (relative residual, orthogonality)
+      void
+      reportResults (const std::string& method,
+                     const Ordinal numCols,
+                     const result_type& result,
+                     const std::string& additionalFieldNames,
+                     const std::string& additionalData,
+                     const bool printFieldNames)
+      {
+        using std::endl;
+
+        const int numProcs = scalarComm_->size();
+        const int myRank = scalarComm_->rank();
+
+        if (myRank == 0)
+          {
+            if (humanReadable_)
+              {
+                out_ << method << " accuracy results:" << endl
+                     << "Scalar type = " << scalarTypeName_ << endl
+                     << "Number of columns = " << numCols << endl
+                     << "Number of (MPI) processes = " << numProcs << endl
+                     << "Absolute residual $\\| A - Q R \\|_2: "
+                     << result[0] << endl
+                     << "Absolute orthogonality $\\| I - Q^* Q \\|_2$: "
+                     << result[1] << endl
+                     << "Test matrix norm $\\| A \\|_F$: "
+                     << result[2] << endl;
+              }
+            else
+              {
+                // Use scientific notation for floating-point numbers
+                out_ << std::scientific;
+
+                if (printFieldNames)
+                  {
+                    out_ << "%method,scalarType,numCols,numProcs"
+                      ",absFrobResid,absFrobOrthog,frobA";
+                    if (! additionalFieldNames.empty())
+                      out_ << "," << additionalFieldNames;
+                    out_ << endl;
+                  }
+
+                out_ << method
+                     << "," << scalarTypeName_
+                     << "," << numCols
+                     << "," << numProcs
+                     << "," << result[0]
+                     << "," << result[1]
+                     << "," << result[2];
+                if (! additionalData.empty())
+                  out_ << "," << additionalData;
+                out_ << endl;
+              }
+          }
+      }
+
+      void
+      testProblem (Matrix< Ordinal, Scalar >& A_local,
+                   Matrix< Ordinal, Scalar >& Q_local,
+                   Matrix< Ordinal, Scalar >& R,
+                   const Ordinal numCols)
+      {
+        const Ordinal numRowsLocal = numCols;
+
+        // A_local: Space for the matrix A to factor -- local to each
+        //   processor.
+        //
+        // A_global: Global matrix (only nonempty on Proc 0); only
+        //   used temporarily.
+        Matrix< Ordinal, Scalar > A_global;
+
+        // This modifies A_local on all procs, and A_global on Proc 0.
+        par_tsqr_test_problem (gen_, A_local, A_global, numCols, scalarComm_);
+
+        if (printMatrices_) {
+          const int myRank = scalarComm_->rank();
+          if (myRank == 0) {
+            err_ << "Input matrix A:" << std::endl;
+          }
+          printGlobalMatrix (err_, A_local, scalarComm_.get(), ordinalComm_.get());
+          if (myRank == 0) {
+            err_ << std::endl;
+          }
+        }
+
+        // Copy the test problem input into R, since the factorization
+        // will overwrite it in place with the final R factor.
+        R.reshape (numCols, numCols);
+        deep_copy (R, Scalar {});
+        deep_copy (R, A_local);
+
+        // Prepare space in which to construct the explicit Q factor
+        // (local component on this processor)
+        Q_local.reshape (numRowsLocal, numCols);
+        deep_copy (Q_local, Scalar {});
+      }
+    };
+
+
+    /// \class DistTsqrBenchmarker
+    /// \brief Generic version of DistTsqr performance test.
+    template< class Ordinal, class Scalar>
+    class DistTsqrBenchmarker {
+      TSQR::Random::NormalGenerator<Ordinal, Scalar> gen_;
+      Teuchos::RCP<MessengerBase<Scalar>> scalarComm_;
+      Teuchos::RCP<MessengerBase<double>> doubleComm_;
+      std::string scalarTypeName_;
+
+      std::ostream& out_;
+      std::ostream& err_;
+      const bool testFactorExplicit_, testFactorImplicit_;
+      const bool humanReadable_, debug_;
+
+    public:
+      using ordinal_type = Ordinal;
+      using scalar_type = Scalar;
+      using magnitude_type =
+        typename Teuchos::ScalarTraits<scalar_type>::magnitudeType;
+      using timer_type = Teuchos::Time;
+
+      /// \brief Constructor, with custom seed value
+      ///
+      /// \param scalarComm [in/out] Communicator object over which
+      ///   to test.
+      /// \param doubleComm [in/out] Communicator object for doubles,
+      ///   used for finding the min and max of timing results over
+      ///   all the MPI processes.
+      /// \param seed [in] 4-element vector; the random seed input of
+      ///   TSQR::Random::NormalGenerator (which see, since there are
+      ///   restrictions on the set of valid seeds)
+      /// \param scalarTypeName [in] Human-readable name of the Scalar
+      ///   template type parameter
+      /// \param out [out] Output stream to which to write results
+      /// \param err [out] Output stream to which to write any
+      ///   debugging outputs (if applicable) or errors
+      /// \param testFactorExplicit [in] Whether to test
+      ///   DistTsqr::factorExplicit()
+      /// \param testFactorImplicit [in] Whether to test
+      ///   DistTsqr::factor() and DistTsqr::explicit_Q()
+      /// \param humanReadable [in] Whether printed results should be
+      ///   easy for humans to read (vs. easy for parsers to parse)
+      /// \param debug [in] Whether to write verbose debug output to
+      ///   err
+      DistTsqrBenchmarker (const Teuchos::RCP< MessengerBase< Scalar > >& scalarComm,
+                           const Teuchos::RCP< MessengerBase< double > >& doubleComm,
+                           const std::vector<int>& seed,
+                           const std::string& scalarTypeName,
+                           std::ostream& out,
+                           std::ostream& err,
+                           const bool testFactorExplicit,
+                           const bool testFactorImplicit,
+                           const bool humanReadable,
+                           const bool debug) :
+        gen_ (seed),
+        scalarComm_ (scalarComm),
+        doubleComm_ (doubleComm),
+        scalarTypeName_ (scalarTypeName),
+        out_ (out),
+        err_ (err),
+        testFactorExplicit_ (testFactorExplicit),
+        testFactorImplicit_ (testFactorImplicit),
+        humanReadable_ (humanReadable),
+        debug_ (debug)
+      {}
+
+      /// \brief Constructor, with default seed value
+      ///
+      /// This constructor sets a default seed (for the pseudorandom
+      /// number generator), which is the same seed (0,0,0,1) each
+      /// time.
+      ///
+      /// \param scalarComm [in/out] Communicator object over which
+      ///   to test.
+      /// \param doubleComm [in/out] Communicator object for doubles,
+      ///   used for finding the min and max of timing results over
+      ///   all the MPI processes.
+      /// \param scalarTypeName [in] Human-readable name of the Scalar
+      ///   template type parameter
+      /// \param out [out] Output stream to which to write results
+      /// \param err [out] Output stream to which to write any
+      ///   debugging outputs (if applicable) or errors
+      /// \param testFactorExplicit [in] Whether to test
+      ///   DistTsqr::factorExplicit()
+      /// \param testFactorImplicit [in] Whether to test
+      ///   DistTsqr::factor() and DistTsqr::explicit_Q()
+      /// \param humanReadable [in] Whether printed results should be
+      ///   easy for humans to read (vs. easy for parsers to parse)
+      /// \param debug [in] Whether to write verbose debug output to
+      ///   err
+      DistTsqrBenchmarker (const Teuchos::RCP< MessengerBase< Scalar > >& scalarComm,
+                           const Teuchos::RCP< MessengerBase< double > >& doubleComm,
+                           const std::string& scalarTypeName,
+                           std::ostream& out,
+                           std::ostream& err,
+                           const bool testFactorExplicit,
+                           const bool testFactorImplicit,
+                           const bool humanReadable,
+                           const bool debug) :
+        scalarComm_ (scalarComm),
+        doubleComm_ (doubleComm),
+        scalarTypeName_ (scalarTypeName),
+        out_ (out),
+        err_ (err),
+        testFactorExplicit_ (testFactorExplicit),
+        testFactorImplicit_ (testFactorImplicit),
+        humanReadable_ (humanReadable),
+        debug_ (debug)
+      {}
+
+      /// \brief Get seed vector for pseudorandom number generator
+      ///
+      /// Fill seed (changing size of vector as necessary) with the
+      /// seed vector used by the pseudorandom number generator.  You
+      /// can use this to resume the pseudorandom number stream from
+      /// where you last were.
+      void
+      getSeed (std::vector<int>& seed) const
+      {
+        gen_.getSeed (seed);
+      }
+
+      /// \brief Run the DistTsqr benchmark
+      ///
+      /// \param numTrials [in] Number of times to repeat the computation
+      ///   in a single timing run
+      /// \param numCols [in] Number of columns in the matrix to test.
+      ///   Number of rows := (# MPI processors) * ncols
+      void
+      benchmark (const int numTrials,
+                 const Ordinal numCols,
+                 const std::string& additionalFieldNames,
+                 const std::string& additionalData,
+                 const bool printFieldNames)
+      {
+        using std::endl;
+
+        // Set up test problem.
+        Matrix< Ordinal, Scalar > A_local, Q_local, R;
+        testProblem (A_local, Q_local, R, numCols);
+
+        // Set up TSQR implementation.
+        DistTsqr<Ordinal, Scalar> par;
+        par.init (scalarComm_);
+
+        // Whether we've printed field names (i.e., column headers)
+        // yet.  Only matters for non-humanReadable output.
+        bool printedFieldNames = false;
+
+        if (testFactorImplicit_)
+          {
+            std::string timerName ("DistTsqr");
+            typedef typename DistTsqr<Ordinal, Scalar>::FactorOutput
+              factor_output_type;
+
+            // Throw away some number of runs, because some MPI libraries
+            // (recent versions of OpenMPI at least) do autotuning for the
+            // first few collectives calls.
+            const int numThrowAwayRuns = 5;
+            for (int runNum = 0; runNum < numThrowAwayRuns; ++runNum)
+              {
+                // Factor the matrix A (copied into R, which will be
+                // overwritten on output)
+                factor_output_type factorOutput = par.factor (R.view());
+                // Compute the explicit Q factor
+                par.explicit_Q (numCols, Q_local.data(), Q_local.stride(1), factorOutput);
+              }
+
+            // Now do the actual timing runs.  Benchmark DistTsqr
+            // (factor() and explicit_Q()) for numTrials trials.
+            timer_type timer (timerName);
+            timer.start();
+            for (int trialNum = 0; trialNum < numTrials; ++trialNum)
+              {
+                // Factor the matrix A (copied into R, which will be
+                // overwritten on output)
+                factor_output_type factorOutput = par.factor (R.view());
+                // Compute the explicit Q factor
+                par.explicit_Q (numCols, Q_local.data(), Q_local.stride(1), factorOutput);
+              }
+            // Cumulative timing on this MPI process.
+            // "Cumulative" means the elapsed time of numTrials executions.
+            const double localCumulativeTiming = timer.stop();
+
+            // reportResults() must be called on all processes, since this
+            // figures out the min and max timings over all processes.
+            reportResults (timerName, numTrials, numCols, localCumulativeTiming,
+                           additionalFieldNames, additionalData,
+                           printFieldNames && (! printedFieldNames));
+            if (printFieldNames && (! printedFieldNames))
+              printedFieldNames = true;
+          }
+
+        if (testFactorExplicit_)
+          {
+            std::string timerName ("DistTsqrRB");
+
+            // Throw away some number of runs, because some MPI libraries
+            // (recent versions of OpenMPI at least) do autotuning for the
+            // first few collectives calls.
+            const int numThrowAwayRuns = 5;
+            for (int runNum = 0; runNum < numThrowAwayRuns; ++runNum)
+              {
+                par.factorExplicit (R.view(), Q_local.view());
+              }
+
+            // Benchmark DistTsqr::factorExplicit() for numTrials trials.
+            timer_type timer (timerName);
+            timer.start();
+            for (int trialNum = 0; trialNum < numTrials; ++trialNum)
+              {
+                par.factorExplicit (R.view(), Q_local.view());
+              }
+            // Cumulative timing on this MPI process.
+            // "Cumulative" means the elapsed time of numTrials executions.
+            const double localCumulativeTiming = timer.stop();
+
+            // Report cumulative (not per-invocation) timing results
+            reportResults (timerName, numTrials, numCols, localCumulativeTiming,
+                           additionalFieldNames, additionalData,
+                           printFieldNames && (! printedFieldNames));
+            if (printFieldNames && (! printedFieldNames))
+              printedFieldNames = true;
+
+            // Per-invocation timings (for factorExplicit() benchmark
+            // only).  localTimings were computed on this MPI process;
+            // globalTimings are statistical summaries of those over
+            // all MPI processes.  We only collect that data for
+            // factorExplicit().
+            std::vector< TimeStats > localTimings;
+            std::vector< TimeStats > globalTimings;
+            par.getFactorExplicitTimings (localTimings);
+            for (std::vector< TimeStats >::size_type k = 0; k < localTimings.size(); ++k)
+              globalTimings.push_back (globalTimeStats (*doubleComm_, localTimings[k]));
+            std::vector< std::string > timingLabels;
+            par.getFactorExplicitTimingLabels (timingLabels);
+
+            if (humanReadable_)
+              out_ << timerName << " per-invocation benchmark results:" << endl;
+
+            const std::string labelLabel ("label,scalarType");
+            for (std::vector< std::string >::size_type k = 0; k < timingLabels.size(); ++k)
+              {
+                // Only print column headers (i.e., field names) once, if at all.
+                const bool printHeaders = (k == 0) && printFieldNames;
+                globalTimings[k].print (out_, humanReadable_,
+                                        timingLabels[k] + "," + scalarTypeName_,
+                                        labelLabel, printHeaders);
+              }
+          }
+      }
+
+    private:
+      /// Report timing results to the given output stream
+      ///
+      /// \param method [in] String to print before reporting results
+      /// \param numTrials [in] Number of times to repeat the computation
+      ///   in a single timing run
+      /// \param numCols [in] Number of columns in the matrix to test.
+      ///   Number of rows := (# MPI processors) * ncols
+      /// \param timing [in] Total benchmark time, as measured on this
+      ///   MPI process.  This may differ on each process; we report
+      ///   the min and the max.
+      ///
+      /// \warning Call on ALL MPI processes, not just Rank 0!
+      void
+      reportResults (const std::string& method,
+                     const int numTrials,
+                     const ordinal_type numCols,
+                     const double localTiming,
+                     const std::string& additionalFieldNames,
+                     const std::string& additionalData,
+                     const bool printFieldNames)
+      {
+        using std::endl;
+
+        // Find min and max timing over all MPI processes
+        TimeStats localStats;
+        localStats.update (localTiming);
+        TimeStats globalStats = globalTimeStats (*doubleComm_, localStats);
+
+        // Only Rank 0 prints the final results.
+        const bool printResults = (doubleComm_->rank() == 0);
+        if (printResults)
+          {
+            const int numProcs = doubleComm_->size();
+            if (humanReadable_)
+              {
+                out_ << method << " cumulative benchmark results (total time over all trials):" << endl
+                     << "Scalar type = " << scalarTypeName_ << endl
+                     << "Number of columns = " << numCols << endl
+                     << "Number of (MPI) processes = " << numProcs << endl
+                     << "Number of trials = " << numTrials << endl
+                     << "Min timing (in seconds) = " << globalStats.min() << endl
+                     << "Mean timing (in seconds) = " << globalStats.mean() << endl
+                     << "Max timing (in seconds) = " << globalStats.max() << endl
+                     << endl;
+              }
+            else
+              {
+                // Use scientific notation for floating-point numbers
+                out_ << std::scientific;
+
+                if (printFieldNames)
+                  {
+                    out_ << "%method,scalarType,numCols,numProcs,numTrials"
+                         << ",minTiming,meanTiming,maxTiming";
+                    if (! additionalFieldNames.empty())
+                      out_ << "," << additionalFieldNames;
+                    out_ << endl;
+                  }
+
+                out_ << method
+                     << "," << scalarTypeName_
+                     << "," << numCols
+                     << "," << numProcs
+                     << "," << numTrials
+                     << "," << globalStats.min()
+                     << "," << globalStats.mean()
+                     << "," << globalStats.max();
+                if (! additionalData.empty())
+                  out_ << "," << additionalData;
+                out_ << endl;
+              }
+          }
+      }
+
+      void
+      testProblem (Matrix< Ordinal, Scalar >& A_local,
+                   Matrix< Ordinal, Scalar >& Q_local,
+                   Matrix< Ordinal, Scalar >& R,
+                   const Ordinal numCols)
+      {
+        const Ordinal numRowsLocal = numCols;
+
+        // A_local: Space for the matrix A to factor -- local to each
+        //   processor.
+        //
+        // A_global: Global matrix (only nonempty on Proc 0); only
+        //   used temporarily.
+        Matrix<Ordinal, Scalar> A_global;
+
+        // This modifies A_local on all procs, and A_global on Proc 0.
+        par_tsqr_test_problem (gen_, A_local, A_global, numCols, scalarComm_);
+
+        // Copy the test problem input into R, since the factorization
+        // will overwrite it in place with the final R factor.
+        R.reshape (numCols, numCols);
+        deep_copy (R, A_local);
+
+        // Prepare space in which to construct the explicit Q factor
+        // (local component on this processor)
+        Q_local.reshape (numRowsLocal, numCols);
+        deep_copy (Q_local, Scalar {});
+      }
+    };
+  } // namespace Test
+} // namespace TSQR
+
+template<class Ordinal, class Scalar>
+class MessengerPairMaker {
+public:
+  using ordinal_type = Ordinal;
+  using scalar_type = Scalar;
+
+  using pair_type = std::pair<
+    Teuchos::RCP<TSQR::MessengerBase<ordinal_type>>,
+    Teuchos::RCP<TSQR::MessengerBase<scalar_type>>
+    >;
+
+  static pair_type
+  makePair (const Teuchos::RCP<const Teuchos::Comm<int>>& comm)
+  {
+    using Teuchos::RCP;
+    using Teuchos::rcp;
+    using Teuchos::rcp_implicit_cast;
+    using TSQR::MessengerBase;
+    using TSQR::TeuchosMessenger;
+
+    auto derivedOrdinalComm =
+      rcp (new TeuchosMessenger<ordinal_type> (comm));
+    auto ordinalComm =
+      rcp_implicit_cast<MessengerBase<ordinal_type> > (derivedOrdinalComm);
+    auto derivedScalarComm =
+      rcp (new TeuchosMessenger<scalar_type> (comm));
+    auto scalarComm =
+      rcp_implicit_cast<MessengerBase<scalar_type>> (derivedScalarComm);
+
+    return {ordinalComm, scalarComm};
+  }
+};
 
 #define TSQR_TEST_DIST_TSQR( ScalarType, typeString )                   \
   do {                                                                  \
-    typedef int ordinal_type;                                           \
-    typedef ScalarType scalar_type;                                     \
-    typedef MessengerPairMaker<ordinal_type, scalar_type>::pair_type pair_type; \
-    typedef DistTsqrVerifier<int, scalar_type> verifier_type;           \
-    \
+    using TSQR::Test::DistTsqrVerifier;                                 \
+    using LO = int;                                                     \
+    using SC = ScalarType;                                              \
+    using verifier_type = DistTsqrVerifier<LO, SC>;                     \
+                                                                        \
     std::string scalarTypeName (typeString);                            \
-    pair_type messPair = MessengerPairMaker< ordinal_type, scalar_type >::makePair (comm); \
+    auto messPair = MessengerPairMaker<LO, SC>::makePair (comm);        \
     verifier_type verifier (messPair.first, messPair.second, seed,      \
-        scalarTypeName, out, err,                       \
-        testFactorExplicit, testFactorImplicit, \
-        humanReadable, printMatrices, debug);           \
+                            scalarTypeName, out, err,                   \
+                            testFactorExplicit, testFactorImplicit,     \
+                            humanReadable, printMatrices, debug);       \
     verifier.verify (numCols, params.additionalFieldNames,              \
-        params.additionalData, params.printFieldNames); \
+                     params.additionalData, params.printFieldNames);    \
     verifier.getSeed (seed);                                            \
-  } while(false)
+  } while (false)
 
 
 #define TSQR_BENCHMARK_DIST_TSQR( theType, typeString )                 \
   do {                                                                  \
-    typedef theType scalar_type;                                                \
-    typedef MessengerBase< scalar_type > base_messenger_type;           \
-    typedef RCP< base_messenger_type > base_messenger_ptr;              \
-    typedef TeuchosMessenger< scalar_type > derived_messenger_type;       \
-    typedef RCP< derived_messenger_type > derived_messenger_ptr;                \
-    typedef DistTsqrBenchmarker<int, scalar_type, timer_type>           \
-    benchmarker_type;                                                   \
-    \
+    using TSQR::Test::DistTsqrBenchmarker;                              \
+    using Teuchos::RCP;                                                 \
+    using SC = theType;                                                 \
+    using base_messenger_type = TSQR::MessengerBase<SC>;                \
+    using base_messenger_ptr = RCP<base_messenger_type>;                \
+    using derived_messenger_type = TSQR::TeuchosMessenger<SC>;          \
+    using derived_messenger_ptr = RCP<derived_messenger_type>;          \
+    using benchmarker_type = DistTsqrBenchmarker<int, SC>;              \
+                                                                        \
     std::string scalarTypeName (typeString);                            \
-    derived_messenger_ptr scalarCommDerived (new derived_messenger_type (comm)); \
-    base_messenger_ptr scalarComm =                                     \
-    rcp_implicit_cast< base_messenger_type > (scalarCommDerived);       \
+    derived_messenger_ptr scalarCommDerived                             \
+      (new derived_messenger_type (comm));                              \
+    auto scalarComm =                                                   \
+      rcp_implicit_cast<base_messenger_type> (scalarCommDerived);       \
     benchmarker_type benchmarker (scalarComm, doubleComm, seed,         \
-        scalarTypeName, out, err,               \
-        testFactorExplicit, testFactorImplicit, \
-        humanReadable, debug);                  \
+                                  scalarTypeName, out, err,             \
+                                  testFactorExplicit,                   \
+                                  testFactorImplicit,                   \
+                                  humanReadable, debug);                \
     benchmarker.benchmark (numTrials, numCols,                          \
-        params.additionalFieldNames,                    \
-        params.additionalData,                          \
-        params.printFieldNames);                        \
+                           params.additionalFieldNames,                 \
+                           params.additionalData,                       \
+                           params.printFieldNames);                     \
     benchmarker.getSeed (seed);                                         \
-  } while(false)
-
-
-////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
+  } while (false)
 
 /// \class DistTsqrTestParameters
 /// \brief Encapsulates values of command-line parameters
-///
 struct DistTsqrTestParameters {
-  DistTsqrTestParameters () :
-    numCols (10),
-    numTrials (10),
-    verify (false),
-    benchmark (false),
-    testReal (true),
-#ifdef HAVE_TPETRATSQR_COMPLEX
-    testComplex (true),
-#endif // HAVE_TPETRATSQR_COMPLEX
-    testFactorExplicit (true),
-    testFactorImplicit (true),
-    printFieldNames (true),
-    printTrilinosTestStuff (true),
-    humanReadable (false),
-    printMatrices (false),
-    debug (false)
-    {}
+  DistTsqrTestParameters () = default;
 
   std::string additionalFieldNames, additionalData;
-  int numCols, numTrials;
-  bool verify, benchmark;
-  bool testReal;
+  int numCols = 10;
+  int numTrials = 10;
+  bool verify = true;
+  bool benchmark = false;
+  bool testReal = true;
 #ifdef HAVE_TPETRATSQR_COMPLEX
-  bool testComplex;
+  bool testComplex = true;
 #endif // HAVE_TPETRATSQR_COMPLEX
-  bool testFactorExplicit, testFactorImplicit;
-  bool printFieldNames, printTrilinosTestStuff;
-  bool humanReadable, printMatrices, debug;
+  bool testFactorExplicit = true;
+  bool testFactorImplicit = true;
+  bool printFieldNames = true;
+  bool printTrilinosTestStuff = true;
+  bool humanReadable = false;
+  bool printMatrices = false;
+  bool debug = false;
 };
 
-  static void
-verify (RCP< const Teuchos::Comm<int> > comm,
-    const DistTsqrTestParameters& params,
-    std::ostream& out,
-    std::ostream& err,
-    std::vector<int>& seed,
-    const bool useSeed)
+static void
+verify (Teuchos::RCP<const Teuchos::Comm<int>> comm,
+        const DistTsqrTestParameters& params,
+        std::ostream& out,
+        std::ostream& err,
+        std::vector<int>& seed,
+        const bool useSeed)
 {
   const bool testReal = params.testReal;
 #ifdef HAVE_TPETRATSQR_COMPLEX
@@ -201,21 +906,18 @@ verify (RCP< const Teuchos::Comm<int> > comm,
   const bool printMatrices = params.printMatrices;
   const bool debug = params.debug;
 
-  if (! useSeed)
-  {
+  if (! useSeed) {
     seed.resize (4);
     seed[0] = 0;
     seed[1] = 0;
     seed[2] = 0;
     seed[3] = 1;
   }
-  if (testReal)
-  {
+  if (testReal) {
     TSQR_TEST_DIST_TSQR( float, "float" );
     TSQR_TEST_DIST_TSQR( double, "double" );
   }
-  if (testComplex)
-  {
+  if (testComplex) {
 #ifdef HAVE_TPETRATSQR_COMPLEX
     using std::complex;
 
@@ -223,22 +925,22 @@ verify (RCP< const Teuchos::Comm<int> > comm,
     TSQR_TEST_DIST_TSQR( complex<double>, "complex<double>" );
 
 #else // Don't HAVE_TPETRATSQR_COMPLEX
-    throw std::logic_error("TSQR was not built with complex "
-        "arithmetic support");
+    throw std::logic_error ("TSQR was not built with complex "
+                            "arithmetic support");
 #endif // HAVE_TPETRATSQR_COMPLEX
   }
 }
 
 
-  static void
-benchmark (RCP< const Teuchos::Comm<int> > comm,
-    const DistTsqrTestParameters& params,
-    std::ostream& out,
-    std::ostream& err,
-    std::vector<int>& seed,
-    const bool useSeed)
+static void
+benchmark (Teuchos::RCP<const Teuchos::Comm<int>> comm,
+           const DistTsqrTestParameters& params,
+           std::ostream& out,
+           std::ostream& err,
+           std::vector<int>& seed,
+           const bool useSeed)
 {
-  typedef Teuchos::Time timer_type;
+  using timer_type = Teuchos::Time;
 
   const bool testReal = params.testReal;
 #ifdef HAVE_TPETRATSQR_COMPLEX
@@ -254,24 +956,26 @@ benchmark (RCP< const Teuchos::Comm<int> > comm,
   const bool humanReadable = params.humanReadable;
   const bool debug = params.debug;
 
-  if (! useSeed)
-  {
+  if (! useSeed) {
     seed.resize (4);
     seed[0] = 0;
     seed[1] = 0;
     seed[2] = 0;
     seed[3] = 1;
   }
-  RCP< MessengerBase< double > > doubleComm =
-    rcp_implicit_cast< MessengerBase< double > > (RCP< TeuchosMessenger< double > > (new TeuchosMessenger< double > (comm)));
+  using Teuchos::rcp;
+  using Teuchos::rcp_implicit_cast;
+  using TSQR::MessengerBase;
+  auto doubleCommSub =
+    rcp (new TSQR::TeuchosMessenger<double> (comm));
+  auto doubleComm =
+    rcp_implicit_cast<MessengerBase<double>> (doubleCommSub);
 
-  if (testReal)
-  {
+  if (testReal) {
     TSQR_BENCHMARK_DIST_TSQR( float, "float" );
     TSQR_BENCHMARK_DIST_TSQR( double, "double" );
   }
-  if (testComplex)
-  {
+  if (testComplex) {
 #ifdef HAVE_TPETRATSQR_COMPLEX
     using std::complex;
 
@@ -279,8 +983,8 @@ benchmark (RCP< const Teuchos::Comm<int> > comm,
     TSQR_BENCHMARK_DIST_TSQR( complex<double>, "complex<double>" );
 
 #else // Don't HAVE_TPETRATSQR_COMPLEX
-    throw std::logic_error("TSQR was not built with complex "
-        "arithmetic support");
+    throw std::logic_error ("TSQR was not built with complex "
+                            "arithmetic support");
 #endif // HAVE_TPETRATSQR_COMPLEX
   }
 }
@@ -296,11 +1000,11 @@ benchmark (RCP< const Teuchos::Comm<int> > comm,
 ///   "help" display (summary of command-line options)
 ///
 /// \return Encapsulation of command-line options
-  static DistTsqrTestParameters
+static DistTsqrTestParameters
 parseOptions (int argc,
-    char* argv[],
-    const bool allowedToPrint,
-    bool& printedHelp)
+              char* argv[],
+              const bool allowedToPrint,
+              bool& printedHelp)
 {
   using std::cerr;
   using std::endl;
@@ -310,8 +1014,9 @@ parseOptions (int argc,
   // Command-line parameters, set to their default values.
   DistTsqrTestParameters params;
   try {
-    Teuchos::CommandLineProcessor cmdLineProc (/* throwExceptions=*/ true,
-        /* recognizeAllOptions=*/ true);
+    using Teuchos::CommandLineProcessor;
+    CommandLineProcessor cmdLineProc (/* throwExceptions=*/ true,
+                                      /* recognizeAllOptions=*/ true);
 
     const char docString[] = "This program tests TSQR::DistTsqr, which "
       "implements the internode-parallel part of TSQR (TSQR::Tsqr).  "
@@ -398,26 +1103,23 @@ parseOptions (int argc,
 
   // Validate command-line options.  We provide default values
   // for unset options, so we don't have to validate those.
-  if (params.numCols <= 0)
+  if (params.numCols <= 0) {
     throw std::invalid_argument ("Number of columns must be positive");
-  else if (params.benchmark && params.numTrials < 1)
+  }
+  else if (params.benchmark && params.numTrials < 1) {
     throw std::invalid_argument ("\"--benchmark\" option requires numTrials >= 1");
+  }
 
   return params;
 }
 
-////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-  int
+int
 main (int argc, char *argv[])
 {
 #ifdef HAVE_MPI
-  typedef RCP< const Teuchos::Comm<int> > comm_ptr;
-
   Teuchos::oblackholestream blackhole;
   Teuchos::GlobalMPISession mpiSession (&argc, &argv, &blackhole);
-  comm_ptr comm = Teuchos::DefaultComm<int>::getComm();
+  auto comm = Teuchos::DefaultComm<int>::getComm();
   const int myRank = comm->getRank();
   // Only Rank 0 gets to write to cout and cerr.  The other MPI
   // process ranks send their output to a "black hole" (something that
@@ -425,9 +1127,7 @@ main (int argc, char *argv[])
   const bool allowedToPrint = (myRank == 0);
   std::ostream& out = allowedToPrint ? std::cout : blackhole;
   std::ostream& err = allowedToPrint ? std::cerr : blackhole;
-
 #else // Don't HAVE_MPI: single-node test
-
   const bool allowedToPrint = true;
   std::ostream& out = std::cout;
   std::ostream& err = std::cerr;
@@ -437,21 +1137,20 @@ main (int argc, char *argv[])
   bool printedHelp = false;
   DistTsqrTestParameters params =
     parseOptions (argc, argv, allowedToPrint, printedHelp);
-  if (printedHelp)
-    return 0;
+  if (printedHelp) {
+    return EXIT_SUCCESS;
+  }
 
   bool success = false;
   bool verbose = false;
   try {
-    if (params.verify)
-    {
+    if (params.verify) {
       std::vector<int> seed(4);
       const bool useSeed = false;
       verify (comm, params, out, err, seed, useSeed);
     }
 
-    if (params.benchmark)
-    {
+    if (params.benchmark) {
       std::vector<int> seed(4);
       const bool useSeed = false;
       benchmark (comm, params, out, err, seed, useSeed);
@@ -459,12 +1158,11 @@ main (int argc, char *argv[])
 
     success = true;
 
-    if (allowedToPrint && params.printTrilinosTestStuff)
+    if (allowedToPrint && params.printTrilinosTestStuff) {
       // The Trilinos test framework expects a message like this.
       out << "\nEnd Result: TEST PASSED" << std::endl;
+    }
   }
   TEUCHOS_STANDARD_CATCH_STATEMENTS(verbose, std::cerr, success);
   return ( success ? EXIT_SUCCESS : EXIT_FAILURE );
 }
-
-
